+++ src/main/java/at/hannibal2/skyhanni/features/garden/farming/GardenCropSpeed.kt
@@ -59,7 +59,6 @@ object GardenCropSpeed {
 
     @SubscribeEvent
     fun onBlockClick(event: CropClickEvent) {
-        if (!GardenAPI.inGarden()) return
         if (event.clickType != ClickType.LEFT_CLICK) return
 
         lastBrokenCrop = event.crop