+++ CHANGELOG.md
@@ -5,6 +5,23 @@
 ### New Features
 + Added Visitor Drop Counter to track all the drops from visitors.
 + Added **Contest Time Needed** - Show the time and missing FF for every crop inside Jacob's Farming Contest inventory.
++ Added **Garden Start Location**
+  + Show the start waypoint for your farm with the currently holding tool.
+  + Auto-detects the start of the farm when farming for the first time
+  + Option to manually set the waypoint with /shcropstartlocation
++ Added **price display next to sack items** (Contributed by HiZe)
+  + Can be disabled
+  + Sortable by price or items stored (both desc/asc)
+  + Option to show prices from Bazaar or NPC
++ Added Farming Fortune Breakdown for Armor and Equipment (Contributed by CalMWolfs)
+  + Works with: Base Stats, Reforge Bonus, Ability Fortune and Green Thumb
++ Added Server Restart Title
++ Added Jacob Contest Stats Summary
+  + Showing Blocks per Second and total Blocks clicked after a farming contest in chat
++ Added City Project Features
+  + Show missing items to contribute inside the inventory
+    + Click on the item name to open the bazaar
+  + Highlight a component in the inventory that can be contributed
 
 ### Changes
 + Added Options for displays Crop Milestone and Best Crop Time.
@@ -16,9 +33,11 @@
   + Improved detection speed
   + Derpy double health support
   + More fair detection logic
++ Added extra setting to allow/block clicks for the 'hide not clickable' feature
 
 ### Fixes
 = Fixed typos in Trevor Trapper texts (Contributed by CalMWolfs)
+= Fixed Hypixel bug that the equipment lore talks about "kills" instead of "visitors" (Contributed by CalMWolfs)
 
 ## Version 0.17 (2023-05-11)